🤪✨refactored profile
