+++ lib/view/profile/schedule.dart
@@ -3,7 +3,7 @@ import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_icons/flutter_icons.dart';
 import 'package:intl/intl.dart';
-import '../widget/widgets.dart';
+import '../../widget/widgets.dart';
 
 class Schedule extends StatefulWidget {
   const Schedule({
