+++ lib/view/views.dart
@@ -5,11 +5,13 @@ export 'auth/forgotPassword.dart';
 export 'mainapp.dart';
 export 'addPost.dart';
 export 'home.dart';
-export 'profilePage.dart';
 export 'Chats/messagePage.dart';
 export 'Chats/chatSchedule.dart';
 export 'notificationsPage.dart';
-export 'schedule.dart';
+export 'profile/profilePage.dart';
+export 'profile/profileFriendsList.dart';
+export 'profile/otherUserProfile.dart';
+export 'profile/schedule.dart';
 export 'drawer/drawer.dart';
 export 'drawer/moreInfo.dart';
 export 'drawer/editProfile.dart';
